+++ santashop-app/src/app/features/pre-registration/confirmation/confirmation.page.ts
@@ -40,6 +40,7 @@ import { manOutline, womanOutline, happyOutline } from 'ionicons/icons';
 import { ChangeDatetimeModalComponent } from './change-datetime-modal/change-datetime-modal.component';
 import { DateTimeSlotsService } from './date-time-slots.service';
 import { firstValueFrom } from 'rxjs';
+import { filter, timeout } from 'rxjs/operators';
 
 @Component({
 	selector: 'app-confirmation',
@@ -154,6 +155,12 @@ export class ConfirmationPage {
 		try {
 			this.analytics.logEvent('cancel_registration');
 			await this.viewService.undoRegistration();
+			await firstValueFrom(
+				this.viewService.registrationComplete$.pipe(
+					filter((isComplete) => !isComplete),
+					timeout({ first: 10000 }),
+				),
+			);
 			this.router.navigate(['/pre-registration/overview']);
 		} catch (error) {
 			await this.errorHandler.handleError(error as IError);